Finish styling of main content
Increased left and right padding of header content to 450px.

Set the border and height for content images. Rounded the border
and removed shrink factor.

Renamed image to content-image in index.html.

Set the width of the caption to the content image width to get
text wrapping and centered the caption text. Also set the caption
text color to a darker gray.

Added a small gap between the caption and the image.

Added a 50px gap between the captioned images.

Added a 50px gap between the heading and list of images.