Finish styling of quote section
Made quote into a flex container and made main axis vertical.
Top and bottom padding but no left and right padding.

Increased the sizing of content images, caption font and width.

Set quote text color, font size and boldness. Also made it italic.
Gave it a width and centered the quote text horizontally.

Gave the quote author text a font size of 18px, boldness of 600.
Set the width to 760 and alignment to flex end to push it to the
right but not so far that its right edge is flush with the right
edge of the quote text.