Finish styling of Call to Action
Made call to action section a flex container and centered the content
box vertically and horizontally. Added padding to get white space
around the content box.

Made the content box a flex container, centered horizontally.
Rounded the box corners, set the background color, width to 1000 px
and added padding.

Consolidated shared properties of header and call to action sign up
buttons: background, color, boldness, rounded corners, padding.

Additionally gave the call to action button a white border and
pushed it to the right edge of the container.

Made the call to action text white and set the font weight of the
call to action title 600.

Fixed typo in call to action text.